+++ projects/rocthrust/thrust/mr/fancy_pointer_resource.h
@@ -68,7 +68,7 @@ public:
    * \param alignment specifies the alignment for the allocation.
    *  Defaults to THRUST_MR_DEFAULT_ALIGNMENT.
    */
-  THRUST_NODISCARD virtual Pointer
+  [[nodiscard]] virtual Pointer
   do_allocate(std::size_t bytes, std::size_t alignment = THRUST_MR_DEFAULT_ALIGNMENT) override
   {
     return static_cast<Pointer>(m_upstream->do_allocate(bytes, alignment));